+++ tinymce-angular-component/src/main/ts/utils/Utils.ts
@@ -8,13 +8,31 @@
 
 import { EventEmitter } from '@angular/core';
 import { EditorComponent } from '../editor/editor.component';
-import { validEvents } from '../editor/Events';
+import { validEvents, Events } from '../editor/Events';
 
 const bindHandlers = (ctx: EditorComponent, editor: any): void => {
-  validEvents.forEach((eventName) => {
+  const allowedEvents = getValidEvents(ctx);
+  allowedEvents.forEach((eventName) => {
     const eventEmitter: EventEmitter<any> = ctx[eventName];
     editor.on(eventName.substring(2), (event: any) => ctx.ngZone.run(() => eventEmitter.emit({ event, editor })));
   });
+};
+
+const getValidEvents = (ctx: EditorComponent): (keyof Events)[] => {
+  const ignoredEvents = parseStringProperty(ctx.ignoreEvents, []);
+  const allowedEvents = parseStringProperty(ctx.allowedEvents, validEvents).filter(
+    (event) => validEvents.includes(event as (keyof Events)) && !ignoredEvents.includes(event)) as (keyof Events)[];
+  return allowedEvents;
+}
+
+const parseStringProperty = (property: string | string[] | undefined, defaultValue: (keyof Events)[]): string[] => {
+  if ( typeof property === 'string') {
+    return property.split(',').map((value) => value.trim());
+  }
+  if ( Array.isArray(property)) {
+    return property;
+  }
+  return defaultValue;
 };
 
 let unique = 0;